+++ Sources/PortalSwift/Utils/Blockchains/PortalBlockchain.swift
@@ -24,6 +24,7 @@ public class PortalBlockchain {
       .eth_signTransaction,
       .eth_signTypedData_v3,
       .eth_signTypedData_v4,
+      .eth_signUserOperation,
       .personal_sign,
       .rawSign
     ],